상품 생성 기능 보완

+++ backend/src/main/java/com/team/shopping/Services/MultiService.java
@@ -22,7 +22,10 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.*;
+
+import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.UUID;
 import java.util.stream.Collectors;
 
 @Service
@@ -38,6 +41,7 @@ public class MultiService {
     private final OptionsService optionsService;
     private final JwtTokenProvider jwtTokenProvider;
     private final FileSystemService fileSystemService;
+    private final TagService tagService;
 
 
     /**
@@ -311,13 +315,19 @@ public class MultiService {
         }
         Category category = this.categoryService.get(requestDTO.getCategoryId());
         Product product = this.productService.save(requestDTO, user, category);
-        String newFile = "/api/product" + "_" + product.getId() + "/";
-        String newUrl = this.fileMove(requestDTO.getUrl(), newFile);
-        if (newUrl != null) {
-            fileSystemService.save(ImageKey.Product.getKey(product.getId().toString()), newUrl);
-            fileSystemService.delete(username);
+        if(requestDTO.getTagList() != null) {
+            for (String tagName : requestDTO.getTagList()) {
+                tagService.save(tagName, product);
+            }
         }
-
+        if(requestDTO.getUrl() != null) {
+            String newFile = "/api/product" + "_" + product.getId() + "/";
+            String newUrl = this.fileMove(requestDTO.getUrl(), newFile);
+            if (newUrl != null) {
+                fileSystemService.save(ImageKey.Product.getKey(product.getId().toString()), newUrl);
+                fileSystemService.delete(username);
+            }
+        }
     }
 
     /**
@@ -422,4 +432,5 @@ public class MultiService {
         return new CategoryResponseDTO(parentCategory.getId(),parentCategory.getName(), childrenDTOList);
     }
 }
+
